pyproject.toml
[project]
name = "foundersNexusBackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=3.2.0",
    "cloudinary==1.44.0",
    "fastapi==0.115.12",
    "motor==3.7.0",
    "pydantic==2.11.3",
    "pyjwt==2.10.1",
    "pymongo==4.12.0",
    "python-dotenv==1.1.0",
    "python-multipart>=0.0.20",
    "requests==2.32.3",
    "tzdata>=2025.2",
    "urllib3==2.4.0",
    "uvicorn>=0.34.1",
]
